Three one-predicate modules become three predicates in the module that asks them
matrixark_mcp_admin was seventeen lines, matrixark_mcp_ingestion eighteen, and the tool-name half
of matrixark_mcp_retrieval a handful: a set of names and a one-line `in` test, each in a file of
its own. Every one was imported by matrixark_mcp_server_request_policy, which is the module that
asks "which class of tool is this", and answering that question in three files meant a fourth
class would plausibly have become a fourth file.

215 production modules become 213.

THE OTHER HALF OF WHY THEY READ AS STRUCTURE

matrixark_mcp_server imported all three and called NONE of them. The import list said the split
was load-bearing and it was not; three dead imports go with the files.

WHAT STAYS

matrixark_mcp_retrieval keeps what is actually retrieval policy --
native_retrieve_fallback_allowed and the three constants it reads -- and loses only the tool-name
test. That function has callers in matrixark_temporal_direct_read,
matrixark_temporal_direct_retrieve and matrixark_mcp_temporal_adapters, so the module is not a
candidate for anything; it was carrying one predicate that belonged elsewhere.

test_python_module_boundaries_part3 follows the predicates rather than losing them: it asked the
three split modules for their predicates and now asks the policy module, which keeps the property
it was written for -- that the entrypoint re-exports rather than growing its own copies. Its
wildcard-import list drops the two removed files and gains matrixark_mcp_server_request_policy,
so the coverage does not shrink with the file count.

Verified: the 60 suites naming the server, the request policy, the module boundaries or the
orphan guard -- 755 tests, 141 failures, the SAME 141 by name on pristine main, none new.
